Add icons and sharing blog articles
Not sure if it's something compatible with dev-standards, but I'm making a Pull Request anyway :)

The idea is that we're already producing content using our blog, and if it's well written it can be as effective as a hand-written "Mode Opératoire". And a blog article is already shared to the community. So instead of rewriting them, let's add them to this repository.

Similarly, an external blog article (external meaning not affiliated with BAM) that is well written could be used as a working standard.