Mo/code quality and component testing (#8)
* Add Setup MO

* Add icons and sharing blog articles

Not sure if it's something compatible with dev-standards, but I'm making a Pull Request anyway :)

The idea is that we're already producing content using our blog, and if it's well written it can be as effective as a hand-written "Mode Opératoire". And a blog article is already shared to the community. So instead of rewriting them, let's add them to this repository.

Similarly, an external blog article (external meaning not affiliated with BAM) that is well written could be used as a working standard.

* Update setup-vscode.mo.md

* add mo on debugging js on device

* Add standard standard

* Add VSCode Jest plugin

* Add mo standard

* Add danger js

* Fix MO template and danger validation for Owner

* Add code-push MO

* Rename file

* Fix standards

* Fix readme link

* Fix owner TBD

* Continue MO

* Turn missing root mentions into warnings

* create code-quality folder and mo for component testing